+++ Idp/Parse.py
@@ -388,6 +388,7 @@ class Theory(object):
         self.vocab_name = kwargs.pop('vocab_name')
         self.constraints = OrderedSet(kwargs.pop('constraints'))
         self.definitions = kwargs.pop('definitions')
+        self.interpretations = {i.name: i for i in kwargs.pop('interpretations')}
 
         self.name = "T" if not self.name else self.name
         self.vocab_name = 'V' if not self.vocab_name else self.vocab_name
@@ -395,7 +396,6 @@ class Theory(object):
         self.clark = {}  # {Declaration: Rule}
         self.def_constraints = {} # {Declaration: Expression}
         self.assignments = Assignments()
-        self.interpretations = {}
 
         for constraint in self.constraints:
             constraint.block = self
@@ -408,9 +408,12 @@ class Theory(object):
 
     def annotate(self, idp):
         assert self.vocab_name in idp.vocabularies, "Unknown vocabulary: " + self.vocab_name
-        voc = idp.vocabularies[self.vocab_name]
+        self.voc = idp.vocabularies[self.vocab_name]
 
-        self.definitions = [e.annotate(self, voc, {}) for e in self.definitions]
+        for i in self.interpretations.values():
+            i.annotate(self) # this updates self.assignments
+
+        self.definitions = [e.annotate(self, self.voc, {}) for e in self.definitions]
         # squash multiple definitions of same symbol declaration
         for d in self.definitions:
             for decl, rule in d.clark.items():
@@ -426,14 +429,14 @@ class Theory(object):
             if type(decl) == SymbolDeclaration and decl.domain:
                 self.def_constraints[decl] = rule.expanded
 
-        self.constraints = OrderedSet([e.annotate(voc, {})        for e in self.constraints])
+        self.constraints = OrderedSet([e.annotate(self.voc, {})        for e in self.constraints])
         self.constraints = OrderedSet([e.expand_quantifiers(self) for e in self.constraints])
 
-        for decl in voc.symbol_decls.values():
+        for decl in self.voc.symbol_decls.values():
             if type(decl) == SymbolDeclaration:
                 self.constraints.extend(decl.typeConstraints)
 
-        for s in voc.terms.values():
+        for s in self.voc.terms.values():
             if not s.code.startswith('_'):
                 self.assignments.assert_(s, None, Status.UNKNOWN, False)
 
@@ -548,10 +551,10 @@ class Rule(object):
             arg, nv = self.args[i],  list(new_vars.values())[i]
             if type(arg) in [Fresh_Variable, Variable] \
             and arg.name in self.vars and arg.name not in new_vars:
-                self.body = self.body.instantiate(arg, nv)
-                self.out = self.out.instantiate(arg, nv) if self.out else self.out
+                self.body = self.body.instantiate(arg, nv, self.block)
+                self.out = self.out.instantiate(arg, nv, self.block) if self.out else self.out
                 for j in range(i, len(self.args)):
-                    self.args[j] = self.args[j].instantiate(arg, nv)
+                    self.args[j] = self.args[j].instantiate(arg, nv, self.block)
             else:
                 eq = AComparison.make('=', [nv, arg])
                 self.body = AConjunction.make('∧', [eq, self.body])
@@ -592,12 +595,12 @@ class Rule(object):
         out = self.body.copy() # in case there is no arguments
         assert len(new_args) == len(self.args) or len(new_args)+1 == len(self.args), "Internal error"
         for old, new in zip(self.args, new_args):
-            out = out.instantiate(old, new)
+            out = out.instantiate(old, new, self.block)
         out = out.expand_quantifiers(theory)
         out = out.interpret(theory)  # add justification recursively
         instance = AppliedSymbol.make(self.symbol, new_args)
         if self.symbol.decl.function:  # a function
-            out = out.instantiate(self.args[-1], instance)
+            out = out.instantiate(self.args[-1], instance, self.block)
         else:
             out = AEquivalence.make('⇔', [instance, out])
         out.block = self.block